requirements.txt
flask==1.1.2
Keras==2.4.3
keras-layer-normalization==0.14.0
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
tensorflow==2.4.0